thêm ngx-pagination

+++ RestaurantManagementISC/RestaurantManagementISC/Models/ThongKeTonKho.cs
@@ -14,9 +14,9 @@ namespace RestaurantManagementISC.Models
         public int id_nguyenlieu { get; set; }
         public string tennguyenlieu { get; set; }
         public string dvt { get; set; }
-        public Nullable<double> tondau { get; set; }
-        public Nullable<double> xuat { get; set; }
-        public Nullable<double> nhap { get; set; }
-        public Nullable<double> toncuoi { get; set; }
+        public Nullable<decimal> tondau { get; set; }
+        public Nullable<decimal> xuat { get; set; }
+        public Nullable<decimal> nhap { get; set; }
+        public Nullable<decimal> toncuoi { get; set; }
     }
 }
